Add grade 8 recipe slot effect and background assets
- Add Item_Recipebg_08.png for grade 8 recipe background sprite
- Add Distortion_RecipeSlotOpen_eff_PL.mat with lavender color for grade 8 exclusive effect
- Register grade 8 entry in UI_RecipeViewData with new background and material
- Fix UpdateCp return type from int? to long? in BattlePreparation

+++ nekoyume/Assets/_Scripts/UI/Widget/BattlePreparation.cs
@@ -289,7 +289,7 @@ namespace Nekoyume.UI
             }
         }
 
-        private int? UpdateCp()
+        private long? UpdateCp()
         {
             switch (_stageType)
             {
